+++ src/main/java/com/shy_polarbear/server/domain/quiz/controller/QuizController.java
@@ -1,6 +1,11 @@
 package com.shy_polarbear.server.domain.quiz.controller;
 
-import com.shy_polarbear.server.domain.quiz.dto.*;
+import com.shy_polarbear.server.domain.quiz.dto.request.MultipleChoiceQuizScoreRequest;
+import com.shy_polarbear.server.domain.quiz.dto.request.OXQuizScoreRequest;
+import com.shy_polarbear.server.domain.quiz.dto.response.MultipleChoiceQuizScoreResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.OXQuizScoreResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.QuizCardResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.WhetherDailyQuizSolvedResponse;
 import com.shy_polarbear.server.domain.quiz.service.QuizService;
 import com.shy_polarbear.server.global.auth.security.PrincipalDetails;
 import com.shy_polarbear.server.global.common.constants.BusinessLogicConstants;
