+++ src/main/java/com/shy_polarbear/server/domain/quiz/dto/response/MultipleChoiceQuizScoreResponse.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.quiz.dto;
+package com.shy_polarbear.server.domain.quiz.dto.response;
 
 import com.shy_polarbear.server.domain.quiz.model.MultipleChoice;
 import com.shy_polarbear.server.domain.quiz.model.MultipleChoiceQuiz;
