+++ src/main/java/com/shy_polarbear/server/domain/quiz/dto/response/WhetherDailyQuizSolvedResponse.java
@@ -1,4 +1,4 @@
-package com.shy_polarbear.server.domain.quiz.dto;
+package com.shy_polarbear.server.domain.quiz.dto.response;
 
 import lombok.Builder;
