+++ src/main/java/com/shy_polarbear/server/domain/quiz/service/QuizService.java
@@ -1,7 +1,12 @@
 package com.shy_polarbear.server.domain.quiz.service;
 
 import com.shy_polarbear.server.domain.point.service.PointService;
-import com.shy_polarbear.server.domain.quiz.dto.*;
+import com.shy_polarbear.server.domain.quiz.dto.request.MultipleChoiceQuizScoreRequest;
+import com.shy_polarbear.server.domain.quiz.dto.request.OXQuizScoreRequest;
+import com.shy_polarbear.server.domain.quiz.dto.response.MultipleChoiceQuizScoreResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.OXQuizScoreResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.QuizCardResponse;
+import com.shy_polarbear.server.domain.quiz.dto.response.WhetherDailyQuizSolvedResponse;
 import com.shy_polarbear.server.domain.quiz.exception.QuizException;
 import com.shy_polarbear.server.domain.quiz.model.*;
 import com.shy_polarbear.server.domain.quiz.repository.*;
